feat: Add Serial Port and Shift Configuration Features
- Introduced serial port configuration with options for enabling/disabling, baud rate, data bits, parity, and stop bits.
- Implemented a new `serialport.h` header for managing UART2 serial communication, including initialization and configuration handling.
- Added shift configuration capabilities, allowing users to define up to three shifts with names, start/end times, and breaks.
- Created `shift_details.h` for managing shift data storage and retrieval from non-volatile storage (NVS).
- Updated web interface to include tabs for Serial Port and Shift configurations, with corresponding HTML and JavaScript for user interaction.
- Enhanced backend API to handle GET and POST requests for serial port and shift configurations, ensuring data persistence and retrieval.

+++ WeighingScale.h
@@ -17,7 +17,7 @@ class WeighingScale {
     WeighingScale(uint8_t dout = HX711_DOUT, uint8_t sck = HX711_SCK, float calib = 1.0f);
 
     bool begin();
-    uint16_t readWeight(bool applyTare = false);
+    float readWeight(bool applyTare = false);
     float readraw();
     void tare();
     void setCalibration(float factor);
